ZREPORT_INTRECTIVE_1001 ABAP report
We created an Interactive Report in SAP ABAP S/4HANA using T-code SE38. In this report, we used two tables: the first is EKKO and the second is EKPO. We defined the structure, work areas, and internal tables for both.

We used a SELECTION SCREEN along with SELECT-OPTIONS. We also implemented various events in the report. For the SELECT-OPTIONS, we wrote logic such that:

If the SELECT-OPTION is blank, the data should still be displayed.

If the SELECT-OPTION contains values, then the corresponding filtered data should be displayed.

In this Interactive Report, we also used the AT LINE-SELECTION event, which is used to redirect from one page to another (i.e., from header to item level details).